refactor(agent-stats): extract h2a module staging
