+++ _extracted/agent-stats-h2a-module/src/agy-chat.ts
