+++ _extracted/agent-stats-h2a-module/src/claude-transcript.ts
