+++ _extracted/agent-stats-h2a-module/src/codex-rollout.ts
