+++ _extracted/agent-stats-h2a-module/src/git-evidence.ts
