+++ _extracted/agent-stats-h2a-module/src/redact.ts
