+++ _extracted/agent-stats-h2a-module/src/report.ts
