+++ _extracted/agent-stats-h2a-module/src/stats.ts
