+++ _extracted/agent-stats-h2a-module/src/track-join.ts
