+++ src/cli.ts
@@ -2612,192 +2612,7 @@ export async function main(): Promise<void> {
     console.log(JSON.stringify(planLifecyclePrune("."), null, 2));
   });
 
-  // WP9 agent-stats: per-agent stats derived from agentic CLI transcripts.
-  // Attribution comes from SESSION EVIDENCE (commit shas the session printed,
-  // h2a registry identity, worktree×branch×time), never from git authorship.
-  function resolveAgentStatsRepoRoot(): string {
-    // Transcripts and the h2a registry key off the MAIN repository checkout, not
-    // a per-worktree path. Derive the main worktree from the common git dir.
-    const commonDir = safeGitRevParse(".", ["--git-common-dir"]);
-    if (commonDir) {
-      const abs = resolve(".", commonDir);
-      // <repo>/.git → <repo>
-      const root = abs.replace(/\/\.git\/?$/, "");
-      if (root && root !== abs) return root;
-      // Bare/linked layouts: fall back to toplevel.
-    }
-    const top = safeGitRevParse(".", ["--show-toplevel"]);
-    return top ?? resolve(".");
-  }
-
-  const agentStats = program
-    .command("agent-stats")
-    .description("Per-agent stats from agentic CLI transcripts (evidence-based attribution, not git authorship)");
-  agentStats
-    .option("--json", "Emit JSON (alias for --format json)")
-    .option("--format <fmt>", "Output format: text | json | md", "text")
-    .action(async (opts) => {
-      const { computeAgentStats, formatStatsTable, buildReport, formatReportMarkdown } = await import(
-        "./agent-stats/index.js"
-      );
-      const repoRoot = resolveAgentStatsRepoRoot();
-      const result = computeAgentStats(repoRoot);
-      const format = opts.json ? "json" : opts.format;
-      if (format === "json") {
-        // Stable schema graphify.agent-stats/v1 — see src/agent-stats/report.ts.
-        console.log(JSON.stringify(buildReport(result), null, 2));
-        return;
-      }
-      if (format === "md") {
-        console.log(formatReportMarkdown(buildReport(result)));
-        return;
-      }
-      console.log(formatStatsTable(result.rows, result.residual, result.conflicts));
-    });
-  agentStats
-    .command("report")
-    .description("Per-agent detail: branches, commits, features, token cost, confidence, anonymized citations")
-    .option("--agent <id>", "Filter by agent id substring")
-    .option("--format <fmt>", "Output format: text | json | md", "text")
-    .action(async (opts) => {
-      const { computeAgentStats, buildReport, filterReportAgents, formatReportMarkdown, formatReportText } =
-        await import("./agent-stats/index.js");
-      const repoRoot = resolveAgentStatsRepoRoot();
-      const report = filterReportAgents(buildReport(computeAgentStats(repoRoot)), opts.agent);
-      if (opts.format === "json") {
-        console.log(JSON.stringify(report, null, 2));
-        return;
-      }
-      if (opts.format === "md") {
-        console.log(formatReportMarkdown(report));
-        return;
-      }
-      console.log(formatReportText(report));
-    });
-  agentStats
-    .command("sync")
-    .description("Parse/refresh transcripts into .graphify/agents/facts.jsonl (incremental)")
-    .option("--full", "Force a full re-parse, ignoring cursors")
-    .action(async (opts) => {
-      const { syncAgentStats } = await import("./agent-stats/index.js");
-      const repoRoot = resolveAgentStatsRepoRoot();
-      const result = syncAgentStats({ repoRoot, full: Boolean(opts.full) });
-      console.log(
-        `agent-stats sync: scanned ${result.scanned} transcripts, parsed ${result.parsed}, ` +
-          `skipped ${result.skipped} (header pre-filter), ${result.inRepo} in-repo; ` +
-          `${result.factsTotal} facts total.`,
-      );
-    });
-  agentStats
-    .command("sessions")
-    .description("List parsed sessions with their evidence-based agent identity")
-    .option("--agent <id>", "Filter by agent id substring")
-    .option("--branch <branch>", "Filter by observed/ground-truth branch")
-    .option("--since <iso>", "Only sessions on/after this ISO date")
-    .option("--json", "Emit JSON (alias for --format json)")
-    .option("--format <fmt>", "Output format: text | json", "text")
-    .action(async (opts) => {
-      const { listSessions, formatSessionsTable, buildSessionsReport } = await import("./agent-stats/index.js");
-      const repoRoot = resolveAgentStatsRepoRoot();
-      const { facts, instances } = listSessions(repoRoot, {
-        agent: opts.agent,
-        branch: opts.branch,
-        since: opts.since,
-      });
-      if (opts.json || opts.format === "json") {
-        // Stable schema graphify.agent-stats.sessions/v1 — see report.ts.
-        console.log(JSON.stringify(buildSessionsReport(facts, instances), null, 2));
-        return;
-      }
-      console.log(formatSessionsTable(facts, instances));
-    });
-  agentStats
-    .command("wp <trackItemId>")
-    .description("Conductor view: agents/sessions joined to a Track work-package (by id or WP label)")
-    .option("--no-pr", "Skip the live `gh` PR-merge attribution step (offline)")
-    .option("--json", "Emit JSON instead of a text view")
-    .action(async (trackItemId, opts) => {
-      const { wpAgentStats, formatWpView } = await import("./agent-stats/index.js");
-      const repoRoot = resolveAgentStatsRepoRoot();
-      const result = wpAgentStats(repoRoot, trackItemId, { skipPrMerges: opts.pr === false });
-      if (opts.json) {
-        console.log(JSON.stringify({
-          item: result.item,
-          links: result.links,
-          sessions: result.sessions.map((s) => ({ factId: s.fact.factId, agentId: s.agentId, rule: s.rule })),
-          evidenced: result.evidenced.map((s) => ({ factId: s.fact.factId, agentId: s.agentId, via: s.via })),
-          mismatch: result.mismatch,
-          rollup: result.rollup,
-        }, null, 2));
-        return;
-      }
-      console.log(formatWpView(result, trackItemId));
-    });
-  agentStats
-    .command("project-graph")
-    .description(
-      "Build a rename-aware graphify graph.json of a project's conversations/sessions (nodes: project/repo/session/agent/branch/commit; reconciles repo renames into ONE project)",
-    )
-    .option(
-      "--config <file>",
-      "JSON ProjectIdentity { canonicalId, label, aliases:[{name,pathPrefixes,remote?}] }. Defaults to the built-in sentropic→graphify lineage.",
-    )
-    .option("--out <path>", "Output graph.json path", ".graphify/project-graph/graph.json")
-    .option("--no-commits", "Omit commit nodes")
-    .option("--no-branches", "Omit branch nodes")
-    .option("--git-since <date>", "Hybrid git skeleton window (git --since); use 'all' for full history", "6 months ago")
-    .option("--git-max-count <n>", "Hybrid git skeleton max commits", "2000")
-    .option("--hub-edges", "Also emit legacy worked-in/conducted-by/belongs-to hub edges")
-    .option("--studio", "Also export a static studio next to the graph.json (graphify studio export)")
-    .action(async (opts) => {
-      const { buildProjectGraphForIdentity } = await import("./agent-stats/index.js");
-      const { readFileSync, mkdirSync, writeFileSync } = await import("node:fs");
-      const { dirname, resolve: pathResolve } = await import("node:path");
-      // Built-in default: the sentropic project and its rename lineage.
-      const defaultIdentity = {
-        canonicalId: "sentropic",
-        label: "Sentropic / Graphify",
-        aliases: [
-          { name: "sentropic", pathPrefixes: ["~/src/sentropic"], remote: "rhanka/sentropic" },
-          { name: "graphify", pathPrefixes: ["~/src/graphify"], remote: "rhanka/graphify" },
-          { name: "regraphify", pathPrefixes: ["/tmp/regraphify", "/tmp/regraphify-brigham"] },
-        ],
-        repoRootForRegistry: resolveAgentStatsRepoRoot(),
-      };
-      const identity = opts.config
-        ? { repoRootForRegistry: resolveAgentStatsRepoRoot(), ...JSON.parse(readFileSync(opts.config, "utf-8")) }
-        : defaultIdentity;
-      const { graph, sessions } = buildProjectGraphForIdentity(identity, {
-        includeCommits: opts.commits !== false,
-        includeBranches: opts.branches !== false,
-        gitSince: opts.gitSince,
-        gitMaxCount: Number.parseInt(opts.gitMaxCount, 10),
-        includeHubEdges: opts.hubEdges === true,
-      });
-      const outPath = pathResolve(opts.out);
-      mkdirSync(dirname(outPath), { recursive: true });
-      writeFileSync(outPath, JSON.stringify(graph, null, 2));
-      console.error(
-        `project-graph: ${graph.nodes.length} nodes, ${graph.links.length} edges from ${sessions} session(s) ` +
-          `across ${identity.aliases.length} rename-alias(es) → ${outPath}`,
-      );
-      if (opts.studio) {
-        try {
-          const { buildStaticStudio } = await import("./studio-export.js");
-          // buildStaticStudio reads <stateDir>/graph.json and writes the studio
-          // into outDir. We keep the export beside the graph.json.
-          const stateDir = dirname(outPath);
-          const studioOut = pathResolve(stateDir, "studio");
-          await buildStaticStudio({ stateDir, outDir: studioOut });
-          console.error(`project-graph: static studio exported to ${studioOut}`);
-        } catch (err) {
-          console.error(
-            `project-graph: studio export skipped (${err instanceof Error ? err.message : String(err)})`,
-          );
-        }
-      }
-    });
-
+  // Agent-stats commands are registered by the extracted h2a module, not Graphify.
   function registerPrCommands(name: "pr" | "prs"): void {
     program.command(`${name} [selector]`)
       .description("Inspect local GitHub pull requests through gh and git worktree data")
