+++ tests/memory-activity-boundary.test.ts
@@ -4,7 +4,8 @@ import { fileURLToPath } from "node:url";
 import { describe, expect, it } from "vitest";
 
 const root = join(dirname(fileURLToPath(import.meta.url)), "..");
-const activityDirectory = join(root, "src", "agent-stats");
+const graphifyActivityDirectory = join(root, "src", "agent-stats");
+const extractedActivityDirectory = join(root, "_extracted", "agent-stats-h2a-module", "src");
 
 function activityFiles(directory: string): string[] {
   if (!existsSync(directory)) return [];
@@ -17,10 +18,12 @@ function activityFiles(directory: string): string[] {
 
 describe("activity evidence boundary", () => {
   it("activity reaches capture only through ActivityEvidenceSource", () => {
-    const directLegacyMemoryEdges = activityFiles(activityDirectory)
-      .filter((file) => /from\s*["']\.\.\/memory-[^"']+["']/.test(readFileSync(file, "utf8")))
+    const rootActivityFiles = activityFiles(graphifyActivityDirectory);
+    const directLegacyMemoryEdges = activityFiles(extractedActivityDirectory)
+      .filter((file) => /from\s*["'][^"']*memory-[^"']+["']/.test(readFileSync(file, "utf8")))
       .map((file) => file.slice(root.length + 1));
 
+    expect(rootActivityFiles, "Graphify must not retain the h2a activity subsystem").toEqual([]);
     expect(
       directLegacyMemoryEdges,
       "activity may reach capture only through ActivityEvidenceSource, never a legacy memory import",
